backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-socketio
gunicorn
psycopg2-binary
python-dotenv
eventlet